Fixed disable items table (scenario config) to not be weird
When changing an item to a different one, it would add a "blank" button
that would only redraw correctly when the window would refresh (such as
from a window resize event).

Also fixed that the unique id for each item in the table could be the
same if choosen from the same choose_elem_button.